modified id pair-structure

+++ rpl-shell/rpl/nodes/verbs.cpp
@@ -22,11 +22,11 @@ verb_node<verb>::verb_node( verb& _verb, const string& id, const std::vector<str
     : _verb(_verb), id(id), index(0), prop(""), parameters(parameters) {}
 
 template <typename verb>
-verb_node<verb>::verb_node( verb& _verb, const pair<string, size_t>& id, const string& prop )
+verb_node<verb>::verb_node( verb& _verb, const pair<string,int>& id, const string& prop )
     : _verb(_verb), id(id.first), index(id.second), prop(prop), parameters({}) {}
 
 template <typename verb>
-verb_node<verb>::verb_node( verb& _verb, const pair<string, size_t>& id, const std::vector<string>& parameters )
+verb_node<verb>::verb_node( verb& _verb, const pair<string,int>& id, const std::vector<string>& parameters )
     : _verb(_verb), id(id.first), index(id.second), prop(""), parameters(parameters) {}
 
 template <typename verb>
@@ -39,7 +39,7 @@ void verb_node<verb>::accept( visitor& v ) {
 show_node::show_node(const string& id, const int& lines, const vector<string>& parameters)
     : verb_node(*this, id, parameters), lines_to_print(lines) {}
 
-show_node::show_node(const pair<string,size_t>& id, const int& lines, const vector<string>& parameters)
+show_node::show_node(const pair<string,int>& id, const int& lines, const vector<string>& parameters)
     : verb_node(*this, id, parameters), lines_to_print(lines) {}
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -52,7 +52,7 @@ set_node::set_node(const string& id, const string& prop, double value)
 ann_node::ann_node(const string& id, const string& prop, double value)
     : verb_node(*this, id, prop), value(value) {}
 
-ann_node::ann_node(const pair<string, size_t>& id, const string& prop, double value)
+ann_node::ann_node(const pair<string,int>& id, const string& prop, double value)
     : verb_node(*this, id, prop), value(value) {}
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -60,7 +60,7 @@ ann_node::ann_node(const pair<string, size_t>& id, const string& prop, double va
 rwr_node::rwr_node(const string& id, const vector<string>& parameters)
     : verb_node(*this, id, parameters) {}
 
-rwr_node::rwr_node(const pair<string,size_t>& id, const vector<string>& parameters)
+rwr_node::rwr_node(const pair<string,int>& id, const vector<string>& parameters)
     : verb_node(*this, id, parameters) {}
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -68,5 +68,5 @@ rwr_node::rwr_node(const pair<string,size_t>& id, const vector<string>& paramete
 opt_node::opt_node(const string& id, const vector<string>& parameters)
     : verb_node(*this, id, parameters) {}
 
-opt_node::opt_node(const pair<string,size_t>& id, const vector<string>& parameters)
+opt_node::opt_node(const pair<string,int>& id, const vector<string>& parameters)
     : verb_node(*this, id, parameters) {}
